Презентер для фильма вынесен в отдельный файл
